Update prm sql1.sql

+++ prm sql1.sql
@@ -5,4 +5,4 @@ create table countries(
 country_id  varchar(10),
 country_name varchar(10),
 check (country_name in ("Italy","India","China")),
-region_id int);
+region_id decimal(10,0));
